Rename dialog classes for readability
