+++ MVVM/Services/Implementations/MoveModsDialogService.cs
@@ -2,22 +2,23 @@
 using Modding_Assistant.MVVM.Services.Interfaces;
 using Modding_Assistant.MVVM.View.Windows;
 using Modding_Assistant.MVVM.ViewModel;
+using Modding_Assistant.MVVM.View.Dialogs;
 
 namespace Modding_Assistant.MVVM.Services.Implementations
 {
-    public class MoveDialogService : IMoveModDialogService
+    public class MoveModsDialogService : IMoveModsDialogService
     {
         private readonly IServiceProvider _serviceProvider;
 
-        public MoveDialogService(IServiceProvider serviceProvider)
+        public MoveModsDialogService(IServiceProvider serviceProvider)
         {
             _serviceProvider = serviceProvider;
         }
 
         public int? ShowNumberDialog()
         {
-            var moveDialog = _serviceProvider.GetRequiredService<MoveModView>();
-            var viewModel = _serviceProvider.GetRequiredService<MoveModViewModel>();
+            var moveDialog = _serviceProvider.GetRequiredService<MoveModsDialog>();
+            var viewModel = _serviceProvider.GetRequiredService<MoveModsViewModel>();
             moveDialog.Owner = _serviceProvider.GetRequiredService<MainWindow>();
             moveDialog.DataContext = viewModel;
             if (moveDialog.ShowDialog() == true)
