+++ MVVM/Services/Interfaces/IMoveModsDialogService.cs
@@ -1,6 +1,6 @@
 ﻿namespace Modding_Assistant.MVVM.Services.Interfaces
 {
-    public interface IMoveModDialogService
+    public interface IMoveModsDialogService
     {
         int? ShowNumberDialog();
     }
