+++ MVVM/View/Dialogs/MoveModsDialog.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Class="Modding_Assistant.MVVM.View.Windows.MoveModView"
+﻿<Window x:Class="Modding_Assistant.MVVM.View.Dialogs.MoveModsDialog"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
